LST 1 funktioniert

+++ Mikroprozesser 22/CMDLine.cs
@@ -9,9 +9,9 @@ namespace Mikroprozesser_22
     public class CommandLine
     {
         public byte counter;
-        public UInt32 command;
+        public UInt16 command;
 
-        public CommandLine(byte cou, UInt32 com)
+        public CommandLine(byte cou, UInt16 com)
         {
             this.counter = cou;
             this.command = com;
